added crud operation
with validation

+++ demo/src/main/java/com/example/demo/service/UserService.java
@@ -1,9 +1,13 @@
 package com.example.demo.service;
 
+import com.example.demo.dto.CreateUserDto;
+import com.example.demo.dto.UpdateUserDto;
 import com.example.demo.dto.UserDto;
 import com.example.demo.entity.User;
 import com.example.demo.repository.UserRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -31,11 +35,62 @@ public class UserService {
         return dto;
     }
 
-    public User addUser(User user) {
+    public User addUser(CreateUserDto dto) {
+        emailExists(dto.getEmail());
+
+        User user = new User();
+        user.setName(dto.getName());
+        user.setEmail(dto.getEmail());
+
         return repository.save(user);
     }
 
     public void deleteUser(Long id) {
+        if (!repository.existsById(id)) {
+            throw new RuntimeException("User not found");
+        }
         repository.deleteById(id);
     }
+
+    public UserDto updateUser(Long id, UpdateUserDto dto) {
+        User user = repository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User not found"));
+
+        if (dto.getName() != null) {
+            user.setName(dto.getName());
+        }
+
+        if (dto.getEmail() != null) {
+            emailExists(dto.getEmail(), id);
+            user.setEmail(dto.getEmail());
+        }
+
+        User saved = repository.save(user);
+        return toDto(saved);
+    }
+
+    void emailExists(String email) {
+        if (repository.existsByEmail(email)) {
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST,
+                    "Email already in use"
+            );
+
+        }
+    }
+
+    void emailExists(String email, Long currentUserId) {
+        boolean emailTaken =
+                repository.existsByEmailAndIdNot(email, currentUserId);
+
+        if (emailTaken) {
+            throw new ResponseStatusException(
+                    HttpStatus.CONFLICT,
+                    "Email already in use"
+            );
+
+        }
+    }
+
+
 }
